+++ app/src/main/java/br/uri/listoflegends/ui/ChampionCard.kt
@@ -1,0 +1,75 @@
+package br.uri.listoflegends.ui
+
+import android.annotation.SuppressLint
+import android.graphics.Bitmap
+import android.util.Log
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardColors
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.asImageBitmap
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.unit.dp
+import br.uri.listoflegends.models.ChampionModel
+import br.uri.listoflegends.services.getImageFromUrl
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+
+@Composable
+fun ChampionCard(champion: ChampionModel) {
+    val coroutineScope = rememberCoroutineScope()
+    var bitmap by remember { mutableStateOf<Bitmap?>(null) }
+
+    LaunchedEffect(champion.icon) {
+        coroutineScope.launch(Dispatchers.IO) {
+            bitmap = getImageFromUrl(champion.icon)
+        }
+    }
+
+    Card(
+        onClick = { /*TODO*/ },
+        modifier = Modifier
+            .padding(8.dp)
+            .size(150.dp)
+    ) {
+        Box(
+            modifier = Modifier.fillMaxSize(),
+            contentAlignment = Alignment.Center
+        ) {
+            bitmap?.let {
+                Image(
+                    bitmap = it.asImageBitmap(),
+                    contentDescription = champion.name,
+                    modifier = Modifier
+                        .size(125.dp)
+                        .padding(bottom = 12.dp) // Add padding here
+                )
+            }
+            Text(
+                text = champion.name,
+                modifier = Modifier.align(Alignment.BottomCenter)
+            )
+        }
+    }
+}
